Changed the way color was determined in bughouse so there is no longer a race condition when a game is being displayed.
Added a PgnHeader for WhiteOnTop to game. This is set to 1 if white is on top.

Added code to respect the flip variable for IcsServers. If you set flip on then the games will show up flipped when viewed.

If flip is not set, and a player you are following is playing a game. You will always view that game from the perspective of that player being on bottom.